fix(actionpack): address Copilot review #2092
- Move Ast class from journey/ast.ts into journey/nodes/node.ts to
  eliminate the circular runtime dep that the re-export introduced.
  Mirrors Rails' file layout (node.rb defines Ast alongside Nodes).
  ast.ts is now a single-line re-export for back-compat.
- Mark the 5 permissions-policy helpers (applyMappings, applyMapping,
  buildDirectives, buildDirective, resolveSource) as private — they
  mirror Rails-private methods.
- Wire Formatter#namedRoutes getter into matchRoute so it's no longer
  dead.
- UploadedFile#open now resets _closed (re-opens after close).
