+++ packages/actionpack/src/action-dispatch/journey/nodes/node.ts
@@ -1,5 +1,4 @@
 import * as Visitors from "../visitors.js";
-export { Ast } from "../ast.js";
 
 export type NodeType = "LITERAL" | "SLASH" | "DOT" | "SYMBOL" | "GROUP" | "STAR" | "CAT" | "OR";
 
@@ -235,5 +234,70 @@ export class Or extends Node {
   }
   override toString(): string {
     return this._children.map((c) => c.toString()).join("|");
+  }
+}
+
+/**
+ * Rails `ActionDispatch::Journey::Ast` (defined alongside Nodes in
+ * `journey/nodes/node.rb`). Walks an AST once and collects symbol nodes,
+ * star nodes, terminals, names, and path-params.
+ */
+export class Ast {
+  readonly tree: Node;
+  readonly pathParams: string[] = [];
+  readonly names: string[] = [];
+  readonly wildcardOptions: Record<string, RegExp> = {};
+  readonly terminals: Node[] = [];
+
+  private readonly symbols: Symbol[] = [];
+  private readonly stars: Star[] = [];
+
+  constructor(tree: Node, formatted: boolean | null | undefined) {
+    this.tree = tree;
+    this.visitTree(formatted);
+  }
+
+  /** Rails alias :root :tree */
+  get root(): Node {
+    return this.tree;
+  }
+
+  set requirements(reqs: Record<string, RegExp>) {
+    for (const node of [...this.symbols, ...this.stars]) {
+      const re = reqs[node.toSym()];
+      if (re) node.regexp = re;
+    }
+  }
+
+  set route(route: unknown) {
+    for (const n of this.terminals) n.memo = route;
+  }
+
+  isGlob(): boolean {
+    return this.stars.length > 0;
+  }
+
+  /** @internal */
+  private visitTree(formatted: boolean | null | undefined): void {
+    for (const node of this.tree) {
+      if (node.isSymbol()) {
+        const sym = node as Symbol;
+        this.pathParams.push(sym.toSym());
+        this.names.push(sym.name);
+        this.symbols.push(sym);
+      } else if (node.isStar()) {
+        const star = node as Star;
+        this.stars.push(star);
+        if (formatted !== false) {
+          const key = star.name;
+          if (!(key in this.wildcardOptions)) {
+            this.wildcardOptions[key] = /.+?/s;
+          }
+        }
+      }
+      if (node.isTerminal()) {
+        this.terminals.push(node);
+      }
+    }
   }
 }
